requirements.txt
fastapi
uvicorn
peewee
python-multipart
pytest
httpx
python-dotenv
jinja2